requirements.txt
RPi.GPIO==0.7.1
paho-mqtt==2.1.0